Added content view, fixed navbar selection, began work on Source Code view
Added the content div that will store the actual contents of the
website. Fixed the navbar to change color when the specific tab is
selected. Added line numbers (fixed amount at the moment) and ready to
add the code (literally a repeat of the current .js file except without
the actual render code cause that would break my brain and project).